backend/requirements.txt
Flask
pymongo
python-dotenv
flask-cors
requests
